[OPENENGSB-2089]moved console classes to internal and removed unused servicesInfo command
